CV-NLP/img_summerization/requirements.txt
Flask==2.3.2
Pillow==10.0.0
pytesseract==0.3.10
transformers==4.30.0
torch==2.0.1
tesseract-ocr

